+++ package.json
@@ -16,11 +16,13 @@
   "license": "MIT",
   "dependencies": {
     "@lit-labs/scoped-registry-mixin": "^1.0.0",
+    "@material/mwc-formfield": "^0.25.3",
     "@material/mwc-list": "^0.25.3",
     "@material/mwc-menu": "^0.25.3",
     "@material/mwc-notched-outline": "^0.25.3",
     "@material/mwc-ripple": "^0.25.3",
     "@material/mwc-select": "^0.25.3",
+    "@material/mwc-switch": "^0.25.3",
     "@material/mwc-textfield": "^0.25.3",
     "custom-card-helpers": "^1.7.2",
     "home-assistant-js-websocket": "^5.11.1",
