+++ rollup.config.dev.js
@@ -4,7 +4,7 @@ import babel from 'rollup-plugin-babel';
 import serve from 'rollup-plugin-serve';
 import { terser } from 'rollup-plugin-terser';
 import json from '@rollup/plugin-json';
-import ignore from './rollup-ignore-plugin.js';
+import ignore from './rollup-plugins/ignore';
 
 export default {
   input: ['src/boilerplate-card.ts'],
