+++ rollup.config.js
@@ -5,7 +5,7 @@ import babel from 'rollup-plugin-babel';
 import { terser } from 'rollup-plugin-terser';
 import serve from 'rollup-plugin-serve';
 import json from '@rollup/plugin-json';
-import ignore from './rollup-ignore-plugin.js';
+import ignore from './rollup-plugins/ignore';
 
 const dev = process.env.ROLLUP_WATCH;
